Update ClientTest to test ClientSocket
The ClientTest is currently using MockClientSocket to meet the test assertions. While MockClientSocket is good for the integration tests in ServerTest, it is not a replacement for actually testing ClientSocket. This commit updates the codebase to allow for testing of ClientSocket which requires dependency injection of the original client, input and output streams.

- Update ClientSocket to take in a clientSocket, input and output streams during initialization. This allows us to pass in a mock input and output during testing.
- Update EchoProtocol to create input and output streams in the run() and pass into ClientSocket as parameters
- Update ClientTest to use ClientSocket instead of MockClientSocket to pass the assertions.